Upgrade the rest client to a class
